Refactored server services to their own modules.
Server services are broken into interfaces.
