+++ services/vfs_service/vfs_service_test.go
@@ -1,23 +1,105 @@
-package services
+package vfs_service
 
 import (
+	"context"
+	"fmt"
 	"path"
 	"testing"
 	"time"
 
 	"github.com/Velocidex/ordereddict"
 	"github.com/alecthomas/assert"
+	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+	"www.velocidex.com/golang/velociraptor/config"
+	config_proto "www.velocidex.com/golang/velociraptor/config/proto"
 	"www.velocidex.com/golang/velociraptor/datastore"
+	"www.velocidex.com/golang/velociraptor/file_store"
+	"www.velocidex.com/golang/velociraptor/file_store/memory"
+	"www.velocidex.com/golang/velociraptor/file_store/test_utils"
 	"www.velocidex.com/golang/velociraptor/flows/proto"
 	flows_proto "www.velocidex.com/golang/velociraptor/flows/proto"
 	"www.velocidex.com/golang/velociraptor/paths"
+	"www.velocidex.com/golang/velociraptor/result_sets"
+	"www.velocidex.com/golang/velociraptor/services"
+	"www.velocidex.com/golang/velociraptor/services/journal"
 	"www.velocidex.com/golang/velociraptor/utils"
 	"www.velocidex.com/golang/velociraptor/vtesting"
 )
 
 type VFSServiceTestSuite struct {
-	BaseServicesTestSuite
+	suite.Suite
+	config_obj *config_proto.Config
+	ctx        context.Context
+	client_id  string
+	flow_id    string
+	sm         *services.Service
+}
+
+func (self *VFSServiceTestSuite) GetMemoryFileStore() *memory.MemoryFileStore {
+	file_store_factory, ok := file_store.GetFileStore(
+		self.config_obj).(*memory.MemoryFileStore)
+	require.True(self.T(), ok)
+
+	return file_store_factory
+}
+
+func (self *VFSServiceTestSuite) SetupTest() {
+	var err error
+	self.config_obj, err = new(config.Loader).WithFileLoader(
+		"../../http_comms/test_data/server.config.yaml").
+		WithRequiredFrontend().WithWriteback().
+		LoadAndValidate()
+	require.NoError(self.T(), err)
+
+	ctx, _ := context.WithTimeout(context.Background(), time.Second*60)
+	self.sm = services.NewServiceManager(ctx, self.config_obj)
+
+	// Start the journaling service manually for tests.
+	self.sm.Start(journal.StartJournalService)
+	self.sm.Start(services.StartNotificationService)
+
+	fmt.Printf("Started VfsService\n")
+	self.sm.Start(StartVFSService)
+
+	self.client_id = "C.12312"
+	self.flow_id = "F.1232"
+}
+
+func (self *VFSServiceTestSuite) TearDownTest() {
+	// Reset the data store.
+	db, err := datastore.GetDB(self.config_obj)
+	require.NoError(self.T(), err)
+
+	db.Close()
+	test_utils.GetMemoryFileStore(self.T(), self.config_obj).Clear()
+	self.sm.Close()
+}
+
+func (self *VFSServiceTestSuite) EmulateCollection(
+	artifact string, rows []*ordereddict.Dict) string {
+
+	// Emulate a Generic.Client.Info collection: First write the
+	// result set, then write the collection context.
+	artifact_path_manager := result_sets.NewArtifactPathManager(
+		self.config_obj, self.client_id, self.flow_id, artifact)
+
+	// Write a result set for this artifact.
+	services.GetJournal().PushRows(artifact_path_manager, rows)
+
+	// Emulate a flow completion message coming from the flow processor.
+	artifact_path_manager = result_sets.NewArtifactPathManager(
+		self.config_obj, "server", "", "System.Flow.Completion")
+
+	services.GetJournal().PushRows(artifact_path_manager,
+		[]*ordereddict.Dict{ordereddict.NewDict().
+			Set("ClientId", self.client_id).
+			Set("FlowId", self.flow_id).
+			Set("Flow", &flows_proto.ArtifactCollectorContext{
+				ClientId:             self.client_id,
+				SessionId:            self.flow_id,
+				ArtifactsWithResults: []string{artifact}})})
+	return self.flow_id
 }
 
 func (self *VFSServiceTestSuite) TestVFSListDirectory() {
@@ -146,5 +228,5 @@ func makeStat(dirname, name string) *ordereddict.Dict {
 }
 
 func TestVFSService(t *testing.T) {
-	suite.Run(t, &VFSServiceTestSuite{BaseServicesTestSuite{}})
+	suite.Run(t, &VFSServiceTestSuite{})
 }
